rename `at/detach_method_relationship.rs`
