+++ bindings/wasm/src/account/update/attach_method_relationships.rs
