+++ bindings/wasm/src/account/update/mod.rs
@@ -1,9 +1,9 @@
 // Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-mod attach_method_relationship;
+mod attach_method_relationships;
 mod create_method;
 mod create_service;
 mod delete_method;
 mod delete_service;
-mod detach_method_relationship;
+mod detach_method_relationships;
